changed salary from decimal to INT

+++ assets/sql/employee.sql
@@ -12,7 +12,7 @@ CREATE TABLE  department(
 CREATE TABLE  role(
   id INT AUTO_INCREMENT,
   title VARCHAR(30) NULL,
-  salary Decimal,
+  salary INT,
   department_id INT,
   PRIMARY KEY (id)
 );
